Refactor S3 client initialization to use environment variables for credentials and update key retrieval method in fetch handler

+++ r2-worker/src/index.ts
@@ -17,20 +17,17 @@ import { post } from './controllers/post';
 
 const BUCKET_NAME = 'profile-images';
 const ACCOUNT_ID = 'd509ce1f3fa7033c5d4169443e122ae1';
-const ACCESS_KEY_ID = '9eebf26d4d4db04a781f7ef984355047';
-const SECRET_ACCESS_KEY = '7a457faa0367c4a0f713753dc52f082e8b9b595fb31ac5d7efe3210d5bc5d376';
-
-const S3 = new S3Client({
-	region: 'auto',
-	endpoint: `https://${ACCOUNT_ID}.r2.cloudflarestorage.com`,
-	credentials: {
-		accessKeyId: ACCESS_KEY_ID,
-		secretAccessKey: SECRET_ACCESS_KEY,
-	},
-});
-
 export default {
 	async fetch(request: Request, env: Env): Promise<Response> {
+		const S3 = new S3Client({
+			region: 'auto',
+			endpoint: `https://${ACCOUNT_ID}.r2.cloudflarestorage.com`,
+			credentials: {
+				accessKeyId: env.ACCESS_KEY_ID,
+				secretAccessKey: env.SECRET_ACCESS_KEY,
+			},
+		});
+
 		const url = new URL(request.url);
 
 		switch (request.method) {
@@ -38,15 +35,26 @@ export default {
 				return post(request, env);
 			}
 			case 'GET': {
-				const key = url.pathname.slice(1);
+				const key = url.searchParams.get('key');
 				console.log({ key });
+
+				if (!key) {
+					return new Response('No key found', {
+						status: 400,
+					});
+				}
 				const signedUrl = await getSignedUrl(S3, new GetObjectCommand({ Bucket: BUCKET_NAME, Key: key }), { expiresIn: 3600 });
 				return new Response(signedUrl, {
 					status: 200,
 				});
 			}
 			case 'DELETE': {
-				const key = url.pathname.slice(1);
+				const key = url.searchParams.get('key');
+				if (!key) {
+					return new Response('No key found', {
+						status: 400,
+					});
+				}
 				await env.MY_BUCKET.delete(key);
 				return new Response('Deleted!');
 			}
